End MIME boundary should be followed by CRLF
The present code mandates that a CRLF is present after intermediate MIME boundary (that starts with --). This change extends that behaviour for end boundary too (that starts and ends with --).
This is useful for processing some bulk emails that uses a single dash character as a boundary and a line full of dashes is present in the content. Gmail and Outlook are able to process such emails, however JavaMail truncates the email after the sequence of five dash characters are consumed from the line full of dash characters.